Turn graphing_inequalities_2 into a proper multiple exercise.
Fixing the validator is another option (adding a null to the array the
guess is compared with is sufficient). However, if the user graphs the
inequality incorrectly and does not select a radio in the point choice,
the exercise would better be graded as incorrect right away. This comes
for free with the multiple-type and would need a more complex validator
otherwise.